Changes to Controllers
Made changes to the Map controller, reseting it to a blank view.
Also messed around with the address field in the WebView.  Ran into
some problems with the address field no resizing correctly.  This is
due to the fact that Apple doesn’t have an address bar built in to the
web view.  So you have to make one custom yourself.

This can be avoided if you do not start off in landscape view and load
the webView.